Update README.md
Conduct a one-sample t-test (IN R)  for income in southern states. Use the t.test( ) in R. Provide the null and the alternative hypothesis for each test. Should this be a one or a two-sided test? Explain/justify your choice. Provide the test results and an interpretation of your results, in light of your choice for a one or two-sided test.